requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.7
Django==3.2.8
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
djangorestframework==3.12.4
idna==3.3
multidict==5.2.0
Pillow==8.4.0
python-telegram-bot-pagination==0.0.2
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.7.0
